br-refactor: redefine profile
Separated account info from profile info.
- Account info: includes credentials for accessing the platform, the
primary email (needed for verification and password recovery), and the
status of the account (verified, active).
- Profile info: includes all the optional information that the user will
be able to provide like the full name, an alternate public email, a
profile picture, a bio, ...

Initially, the profile info will be loaded in a table but it could
change in the future to allow easier adding of new attributes.